+++ packages/learn-card-cli/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @learncard/cli
 
+## 3.3.82
+
+### Patch Changes
+
+-   Updated dependencies [[`a945a7086a0848b04f9f2c8b1588e9e3826edeff`](https://github.com/learningeconomy/LearnCard/commit/a945a7086a0848b04f9f2c8b1588e9e3826edeff), [`a945a7086a0848b04f9f2c8b1588e9e3826edeff`](https://github.com/learningeconomy/LearnCard/commit/a945a7086a0848b04f9f2c8b1588e9e3826edeff)]:
+    -   @learncard/core@9.3.20
+    -   @learncard/init@2.0.0
+    -   @learncard/didkit-plugin@1.5.6
+    -   @learncard/learn-cloud-plugin@2.1.15
+
+## 3.3.81
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.3.40
+
+## 3.3.80
+
+### Patch Changes
+
+-   Updated dependencies [[`099f06d92d77a43fe953c0d63767355ff050ef56`](https://github.com/learningeconomy/LearnCard/commit/099f06d92d77a43fe953c0d63767355ff050ef56)]:
+    -   @learncard/didkit-plugin@1.5.5
+    -   @learncard/init@1.3.39
+    -   @learncard/learn-cloud-plugin@2.1.14
+
+## 3.3.79
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.3.38
+
 ## 3.3.78
 
 ### Patch Changes