+++ packages/learn-card-network/simple-signing-client/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @welibraryos/lca-api-client
 
+## 1.0.3
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-service@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-service@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes