+++ packages/plugins/simple-signing-plugin/CHANGELOG.md
@@ -1,1 +1,17 @@
 # lca-api-plugin
+
+## 1.0.3
+
+### Patch Changes
+
+-   [#635](https://github.com/learningeconomy/LearnCard/pull/635) [`a945a7086a0848b04f9f2c8b1588e9e3826edeff`](https://github.com/learningeconomy/LearnCard/commit/a945a7086a0848b04f9f2c8b1588e9e3826edeff) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Remove dependency on ceramic
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/simple-signing-client@1.0.2